Reactoring complet des Items sur le back

+++ back/src/main/java/com/g1b/station_back/dto/TransactionLineRequestDTO.java
@@ -5,7 +5,7 @@ import jakarta.validation.constraints.NotNull;
 
 public record TransactionLineRequestDTO(
         @NotNull(message = "Item ID cannot be null")
-        Long idItem,
+        Integer idItem,
         @NotNull(message = "Quantity cannot be null")
         @Min(value = 1, message = "Quantity must be at least 1")
         Integer quantity
